+++ packages/ui/components/CssVariables/transformColors.js
@@ -1,11 +1,13 @@
-export default function transformColors ({ colors, palette }) {
+export default function transformColors ({ colors = {}, palette = {} }) {
   const res = Object.keys(colors).reduce((vars, color) => {
-    vars[`--color-${color}`] = colors[color]
+    const prefix = color.includes('--color') ? '' : '--color-'
+    vars[`${prefix}${color}`] = colors[color]
     return vars
   }, {})
 
   return Object.keys(palette).reduce((vars, color) => {
-    vars[`--palette-${color}`] = palette[color]
+    const prefix = color.includes('--palette') ? '' : '--palette-'
+    vars[`${prefix}${color}`] = palette[color]
     return vars
   }, res)
 }